naegginiggi/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
daphne==3.0.2
Django==4.1.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
mysqlclient==2.1.1
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyMySQL==1.0.3
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==1.26.14
zope.interface==6.0
